Created a new page solely for analysis; Removed analysis from Galaga page; Separated controls from Teammate Video; Analysis page displays Teammate Video front and center
